requirements.txt
matplotlib==3.10.8
numpy==2.4.3
pandas==3.0.2
Pillow==12.1.1
reportlab==4.4.10
scikit-learn==1.8.0
torch==2.11.0
torchvision==0.26.0
